Move ONNX to class
